optimize netcode. FE part

+++ service/src/main/java/com/metalheart/service/state/GameStateService.java
@@ -10,5 +10,5 @@ public interface GameStateService {
 
     void unregisterPlayer(String playerId);
 
-    State step(Integer tickDelay);
+    State step(float tickDelay);
 }
